requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
pandas==2.1.3
dash==2.14.1
yfinance==0.2.31
python-dotenv==1.0.0
webdriver-manager==4.0.1
plotly==5.18.0
aiohttp==3.9.1
pillow==10.1.0
dash-bootstrap-components[icons]==1.6.0
gunicorn==21.2.0
